% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_GRanges.r
\docType{data}
\name{example_GRanges}
\alias{example_GRanges}
\title{Toy example of GRanges object}
\format{
A \code{GRanges} object with 11 rows and several columns:
  \describe{
    \item{\code{seqnames}}{Chromosome number (only \code{Y} in this example)}
    \item{\code{ranges}}{\code{IRanges} object for which \code{start}=\code{end}: position on the chromosome}
    \item{\code{reference}}{Referent nucleotid (\code{A}, \code{C}, \code{G} or \code{T})}
    \item{\code{alteration}}{Alteration nucleotid (\code{A}, \code{C}, \code{G} or \code{T})}
  }
}
\description{
Toy example of GRanges object
}
\keyword{data}
