% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_list_Y.r
\docType{data}
\name{variant_list_Y}
\alias{variant_list_Y}
\title{Toy example of data.frame}
\format{
A data frame with 5 observations on the following 5 variables:
  \describe{
    \item{\code{chr}}{Chromosome number (only \code{Y} in this example)}
    \item{\code{pos}}{Position on the chromosome (numeric)}
    \item{\code{ref}}{Referent nucleotid (\code{A}, \code{C}, \code{G} or \code{T})}
    \item{\code{alt}}{Alteration nucleotid (\code{A}, \code{C}, \code{G} or \code{T})}
    \item{\code{uniprot}}{Uniprot protein identifier (factor)}
  }
}
\description{
Toy example of data.frame
}
\keyword{data}
